save non numeric sequence ids
this allows users to save string values as a sequence ids. it's
assumed that the values provided will sort lexicographically as they
represent some sort of sequence.

prior to this non numeric sequence ids were skipped in the frequency
check. There was no api to know this was happening but we never told
anyone they could save numbers.

this may be a major bump because anyone who currently has a bug where
they pass a non numeric value as a sequence id will start saving that
value instead of ignoring it.